fix: Remove duplicate files and rename activity folder to view
