+++ app/src/androidTest/java/campus/tech/kakao/map/DataSearchActivityTest.kt
@@ -7,7 +7,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
-import campus.tech.kakao.map.activity.DataSearchActivity
+import campus.tech.kakao.map.view.DataSearchActivity
 import org.junit.Rule
 import org.junit.Test
