+++ app/src/main/AndroidManifest.xml
@@ -16,10 +16,10 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".activity.MapErrorActivity"
+            android:name=".view.MapErrorActivity"
             android:exported="true" />
         <activity
-            android:name=".activity.HomeMapActivity"
+            android:name=".view.HomeMapActivity"
             android:allowBackup="true"
             android:exported="true">
             <intent-filter>
@@ -29,7 +29,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activity.DataSearchActivity"
+            android:name=".view.DataSearchActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize" />
     </application>
