+++ app/src/main/res/layout/activity_map_error.xml
@@ -12,7 +12,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.MapErrorActivity">
+        tools:context=".view.MapErrorActivity">
 
         <TextView
             android:id="@+id/error_notice"
